feat: extract Postdata handler
